- created configuration model - removed parmaeters from behaviour model - added item for configurations in newResourceContextMenu
