+++ alica-plan-designer-fx-modelmanagement/src/main/java/de/unikassel/vs/alica/planDesigner/modelmanagement/ModelManager.java
@@ -712,12 +712,12 @@ public class ModelManager implements Observer {
 
     public void changeAttribute(PlanElement planElement, String elementType, String attribute, Object newValue, Object oldValue) {
         try {
-            if (planElement instanceof Behaviour && attribute.equals("parameters")) {
-                Behaviour behaviour = (Behaviour) planElement;
-                behaviour.modifyParameter((Map.Entry<String, String>) newValue, (Map.Entry<String, String>) oldValue);
-            } else {
+//            if (planElement instanceof Behaviour && attribute.equals("parameters")) {
+//                Behaviour behaviour = (Behaviour) planElement;
+//                behaviour.modifyParameter((Map.Entry<String, String>) newValue, (Map.Entry<String, String>) oldValue);
+//            } else {
                 BeanUtils.setProperty(planElement, attribute, newValue);
-            }
+//            }
 
             if (attribute.equals("name")) {
                 String ending = "";
