+++ alica-plan-designer-fx-view/src/main/resources/pld_globals.properties
@@ -56,6 +56,7 @@ label.menu.new.plantype=PlanType
 label.menu.new.roleset=RoleSet
 label.menu.new.folder=Folder
 label.menu.new.taskrepository=TaskRepository
+label.menu.new.configuration=Configuration
 label.choose.parentDirectory=Choose parent directory
 label.choose.task=Choose task
 label.generation.completed=Generation completed!
